split up integration tests into multiple files and removed test suite, closes #24
