+++ 52n-wps-webapp/src/test/java/org/n52/wps/test/AllTestsIT.java
@@ -19,8 +19,6 @@ import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.parsers.ParserConfigurationException;
 
 import org.geotools.coverage.grid.GridCoverage2D;
-import org.junit.runner.RunWith;
-import org.junit.runners.Suite;
 import org.n52.wps.io.data.IData;
 import org.n52.wps.io.datahandler.parser.GeotiffParser;
 import org.w3c.dom.Document;
@@ -29,11 +27,8 @@ import org.w3c.dom.NodeList;
 import org.xml.sax.InputSource;
 import org.xml.sax.SAXException;
 
-@RunWith(Suite.class)
-@Suite.SuiteClasses({GetCapabilitiesKVPTests.class, GetCapabilitiesPOSTTester.class, ExecutePOSTTester.class, DescribeProcessKVPTester.class, DescribeProcessPOSTTester.class})
 public class AllTestsIT {
 
-    // suite.addTestSuite(DescribeProcessKVPTester.class); // test
     public static int getPort() {
         return Integer.parseInt(System.getProperty("test.port", "8080"));
     }
@@ -65,7 +60,9 @@ public class AllTestsIT {
         assertThat(referencedDocument, referencedDocument, containsString("ExecuteResponse"));
         assertThat(referencedDocument,
                    referencedDocument,
-                   anyOf(containsString("AAEGAAMAAAABAAEAAAEVAAMAAAABA"), containsString("Tk9SVEg6IDIyOD"), containsString("SUkqAAgAAAASAAABAwABAAAAIwAA")));
+                   anyOf(containsString("AAEGAAMAAAABAAEAAAEVAAMAAAABA"),
+                         containsString("Tk9SVEg6IDIyOD"),
+                         containsString("SUkqAAgAAAASAAABAwABAAAAIwAA")));
     }
 
     public static String getRefAsString(String response) throws ParserConfigurationException, SAXException, IOException {
